pr 반영

+++ src/main/java/com/example/easytable/repository/restaurant/RestaurantRepository.java
@@ -4,6 +4,9 @@ import com.example.easytable.entity.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface RestaurantRepository extends JpaRepository<Restaurant, Integer>, RestaurantRepositoryCustom {
+public interface RestaurantRepository extends JpaRepository<Restaurant, Long>, RestaurantRepositoryCustom {
+
+
 }
